CI: cron run for flakey tests
It's possible for some tests to pass due to us getting lucky with
iteration order. Experimentally it seems about 20 runs guarantees us a
failure, so with about a 100 it's bound to blow up. This tends to only
be an issue with implementations where the JSON diff trips up due to
language-tagged values not being sorted by language tag.