feat: add application-defined context validation
A context validation function can now be set on the processor which will
be used to check the context. If this function returns false, an invalid
context error is raised. The function can be used to apply additional
constraints to the context, or ensure it is in a known/correct state.