feat: initialise context with preprocessed context
When we're processing many documents which all share a remote context,
processing the same context again and again is a waste of resources.
Instead, the context can be preprocessed and cloned each time we need
it. This only works for the first context if it's a remote context, as
processing other contexts might be influenced by existing context state.

This can easily slash allocs by an order of magnitude.