perf: speed up compaction by tracking prefixes
Before we had to iterate through all definitions to find prefixes. But
in general there are far far far fewer prefixes than there are terms. In
the AS context there are 140 or so terms, and 4 prefixes.

Track the prefixes separately so we can do a quick lookup for those when
needed.